liana-ui: move receive modal to their own submodule
